Add files via upload
unittestbug4.c initializes the game, sets the supply count of the last two cards in the CARD enum to zero, and calls isGameOver(). This call should return 0 because the end-game state (where the supply counts of 3 cards that are not Province cards are zero) has not been reached. Then another card's supply count is set to zero and isGameOver() is called again. This function should return 1 because the end-game state has been met, but the function call returns 0.